refactor : layout 수정 및 post 경로 이동
