+++ app/(root)/(menubar-exist)/post/[postId]/components/BotContainer.tsx
