+++ app/(root)/(menubar-exist)/post/[postId]/components/icons/ThumbsUpIcon.tsx
