+++ app/(root)/(menubar-exist)/post/[postId]/components/mid/ThumbsUpCount.tsx
