+++ app/(root)/(menubar-exist)/post/[postId]/page.tsx
@@ -3,9 +3,9 @@
 import React, { useState, useCallback } from 'react';
 import TopContainer from './components/TopContainer';
 import MidContainer from './components/MidContainer';
-import usePostDetail from '../../../../hooks/usePostDetail';
+import usePostDetail from '../../../../../hooks/usePostDetail';
 import BotContainer from './components/BotContainer';
-import useComments from '../../../../hooks/useComments';
+import useComments from '../../../../../hooks/useComments';
 
 export default function PostDetailPage() {
   const userDetail = usePostDetail();
