+++ app/(root)/(menubar-exist)/post/edit/[postid]/page.tsx
@@ -1,14 +1,14 @@
 'use client';
 
 import { FormEvent, SetStateAction, useEffect, useState } from 'react';
-import ImageUploadContainer from '../../../../../components/post/ImageUploadContainer';
-import ContentContainer from '../../../../../components/post/ContentContainer';
-import HashTagContainer from '../../../../../components/post/HashTagContainer';
-import PrivacySettingContainer from '../../../../../components/post/PrivacySettingContainer';
+import ImageUploadContainer from '../../components/ImageUploadContainer';
+import ContentContainer from '../../components/ContentContainer';
+import HashTagContainer from '../../components/HashTagContainer';
+import PrivacySettingContainer from '../../components/PrivacySettingContainer';
 import { useParams, useRouter } from 'next/navigation';
-import { getPostEditData } from '../../../../../api/post/getPostEditData';
-import { postEdit } from '../../../../../api/post/postEdit';
-import MyPageModal from '../../../../../components/modal/common/postAlarmModal/PostAlarmModal';
+import { getPostEditData } from '../../../../../../api/post/getPostEditData';
+import { postEdit } from '../../../../../../api/post/postEdit'; 
+import MyPageModal from '../../../../../../components/modal/common/postAlarmModal/PostAlarmModal';
 
 type ImageData = {
   fileName: string;
